feat(JobsFields): New fields for controllers.

+++ src/main/java/mypackage/hrms/dataAccess/abstracts/JobTitlesDao.java
@@ -1,11 +1,10 @@
 package mypackage.hrms.dataAccess.abstracts;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import mypackage.hrms.entities.concretes.JobTitles;
 
 import java.util.Optional;
 
-public interface JobTitlesDao extends JpaRepository<JobTitles, Integer> {
+public interface JobTitlesDao extends JpaRepository<JobTitles, Long> {
     Optional<JobTitles> findByTitle(String title);
 }
